backend/eve/requirements.txt
django
channels
channels_redis
asgi_redis
google-cloud-translate